Move functions from harmony_regridder.py into new separated modules.
